refactor: Move db config to database dir
