+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { APP_FILTER } from '@nestjs/core';
 import i18n from 'i18n';
 
 import { AllExceptionsFilter } from '@app/exceptions/http-exception.filter';
-import Database from '@app/config/database';
+import Database from '@database/connect';
 
 /**
  * Import Middleware
